+++ src/enc/encoder.rs
@@ -11,11 +11,11 @@ use crate::{
 };
 use core::ops::{Deref, DerefMut};
 
-const LZMA2_UNCOMPRESSED_LIMIT: u32 = (2 << 20) - MATCH_LEN_MAX as u32;
-const LZMA2_COMPRESSED_LIMIT: u32 = (64 << 10) - 26;
+const LZMA2_UNCOMPRESSED_LIMIT: u64 = (2 << 20) - MATCH_LEN_MAX as u64;
+const LZMA2_COMPRESSED_LIMIT: u64 = (64 << 10) - 26;
 
-const DIST_PRICE_UPDATE_INTERVAL: u32 = FULL_DISTANCES as u32;
-const ALIGN_PRICE_UPDATE_INTERVAL: u32 = ALIGN_SIZE as u32;
+const DIST_PRICE_UPDATE_INTERVAL: u64 = FULL_DISTANCES as u64;
+const ALIGN_PRICE_UPDATE_INTERVAL: u64 = ALIGN_SIZE as u64;
 const PRICE_UPDATE_INTERVAL: usize = 32;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum EncodeMode {
@@ -23,7 +23,7 @@ pub enum EncodeMode {
     Normal,
 }
 pub trait LZMAEncoderTrait {
-    fn get_next_symbol(&mut self, encoder: &mut LZMAEncoder) -> u32;
+    fn get_next_symbol(&mut self, encoder: &mut LZMAEncoder) -> u64;
     fn reset(&mut self) {}
 }
 
@@ -32,7 +32,7 @@ pub enum LZMAEncoderModes {
     Normal(NormalEncoderMode),
 }
 impl LZMAEncoderTrait for LZMAEncoderModes {
-    fn get_next_symbol(&mut self, encoder: &mut LZMAEncoder) -> u32 {
+    fn get_next_symbol(&mut self, encoder: &mut LZMAEncoder) -> u64 {
         match self {
             LZMAEncoderModes::Fast(a) => a.get_next_symbol(encoder),
             LZMAEncoderModes::Normal(a) => a.get_next_symbol(encoder),
@@ -58,20 +58,20 @@ pub struct LZMAEncoder {
 
 pub(super) struct LZMAEncData {
     pub(super) nice_len: usize,
-    dist_price_count: i32,
-    align_price_count: i32,
-    dist_slot_prices_size: u32,
-    dist_slot_prices: crate::Vec<crate::Vec<u32>>,
-    full_dist_prices: [[u32; FULL_DISTANCES]; DIST_STATES],
-    align_prices: [u32; ALIGN_SIZE],
-    pub(super) back: i32,
-    pub(super) read_ahead: i32,
-    pub(super) uncompressed_size: u32,
+    dist_price_count: i64,
+    align_price_count: i64,
+    dist_slot_prices_size: u64,
+    dist_slot_prices: crate::Vec<crate::Vec<u64>>,
+    full_dist_prices: [[u64; FULL_DISTANCES]; DIST_STATES],
+    align_prices: [u64; ALIGN_SIZE],
+    pub(super) back: i64,
+    pub(super) read_ahead: i64,
+    pub(super) uncompressed_size: u64,
 }
 
 impl LZMAEncoder {
-    pub fn get_dist_slot(dist: u32) -> u32 {
-        if dist <= DIST_MODEL_START as u32 {
+    pub fn get_dist_slot(dist: u64) -> u64 {
+        if dist <= DIST_MODEL_START as u64 {
             return dist;
         }
         let mut n = dist;
@@ -106,10 +106,10 @@ impl LZMAEncoder {
 
     pub fn get_mem_usage(
         mode: EncodeMode,
-        dict_size: u32,
-        extra_size_before: u32,
+        dict_size: u64,
+        extra_size_before: u64,
         mf: MFType,
-    ) -> u32 {
+    ) -> u64 {
         let mut m = 80;
         match mode {
             EncodeMode::Fast => {
@@ -127,12 +127,12 @@ impl LZMAEncoder {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         mode: EncodeMode,
-        lc: u32,
-        lp: u32,
-        pb: u32,
+        lc: u64,
+        lp: u64,
+        pb: u64,
         mf: MFType,
-        depth_limit: i32,
-        dict_size: u32,
+        depth_limit: i64,
+        dict_size: u64,
         nice_len: usize,
     ) -> (Self, LZMAEncoderModes) {
         let fast_mode = mode == EncodeMode::Fast;
@@ -206,7 +206,7 @@ impl LZMAEncoder {
         self.rep_len_encoder.reset();
         self.data.dist_price_count = 0;
         self.data.align_price_count = 0;
-        self.data.uncompressed_size += (self.data.read_ahead + 1) as u32;
+        self.data.uncompressed_size += (self.data.read_ahead + 1) as u64;
         self.data.read_ahead = -1;
         mode.reset();
     }
@@ -235,14 +235,14 @@ impl LZMAEncoder {
         &mut self,
         rc: &mut RangeEncoder<W>,
     ) -> crate::io::write_result!(W, ()) {
-        let pos_state = (self.lz.get_pos() - self.data.read_ahead) as u32 & self.coder.pos_mask;
+        let pos_state = (self.lz.get_pos() - self.data.read_ahead) as u64 & self.coder.pos_mask;
         rc.encode_bit(
             &mut self.coder.is_match[self.coder.state.get() as usize],
             pos_state as usize,
             1,
         )?;
         rc.encode_bit(&mut self.coder.is_rep, self.coder.state.get() as usize, 0)?;
-        self.encode_match(u32::MAX, MATCH_LEN_MIN as u32, pos_state, rc)?;
+        self.encode_match(u64::MAX, MATCH_LEN_MIN as u64, pos_state, rc)?;
         Ok(())
     }
 
@@ -287,7 +287,7 @@ impl LZMAEncoder {
         let len = mode.get_next_symbol(self);
 
         assert!(self.data.read_ahead >= 0);
-        let pos_state = (self.lz.get_pos() - self.data.read_ahead) as u32 & self.pos_mask;
+        let pos_state = (self.lz.get_pos() - self.data.read_ahead) as u64 & self.pos_mask;
 
         if self.data.back == -1 {
             assert!(len == 1);
@@ -298,11 +298,11 @@ impl LZMAEncoder {
         } else {
             let state = self.state.get() as usize;
             rc.encode_bit(&mut self.is_match[state], pos_state as usize, 1)?;
-            if self.data.back < REPS as i32 {
+            if self.data.back < REPS as i64 {
                 let match_len2 = self.lz.get_match_len2(
                     -self.data.read_ahead,
                     self.reps[self.data.back as usize],
-                    len as i32,
+                    len as i64,
                 );
 
                 let start = (self.lz.read_pos - 20).max(0) as usize;
@@ -319,29 +319,29 @@ impl LZMAEncoder {
                 );
                 let state = self.state.get() as usize;
                 rc.encode_bit(&mut self.is_rep, state, 1)?;
-                self.encode_rep_match(self.data.back as u32, len, pos_state, rc)?;
+                self.encode_rep_match(self.data.back as u64, len, pos_state, rc)?;
             } else {
                 let match_len2 = self.lz.get_match_len2(
                     -self.data.read_ahead,
-                    self.data.back - REPS as i32,
-                    len as i32,
+                    self.data.back - REPS as i64,
+                    len as i64,
                 );
                 assert_eq!(match_len2, len);
                 let state = self.state.get() as usize;
                 rc.encode_bit(&mut self.is_rep, state, 0)?;
-                self.encode_match((self.data.back - REPS as i32) as u32, len, pos_state, rc)?;
+                self.encode_match((self.data.back - REPS as i64) as u64, len, pos_state, rc)?;
             }
         }
-        self.data.read_ahead -= len as i32;
+        self.data.read_ahead -= len as i64;
         self.data.uncompressed_size += len;
         Ok(true)
     }
 
     fn encode_match<W: Write>(
         &mut self,
-        dist: u32,
-        len: u32,
-        pos_state: u32,
+        dist: u64,
+        len: u64,
+        pos_state: u64,
         rc: &mut RangeEncoder<W>,
     ) -> crate::io::write_result!(W, ()) {
         self.state.update_match();
@@ -365,14 +365,14 @@ impl LZMAEncoder {
             let base = (2 | (dist_slot & 1)) << footer_bits;
             let dist_reduced = dist - base;
 
-            if dist_slot < DIST_MODEL_END as u32 {
+            if dist_slot < DIST_MODEL_END as u64 {
                 rc.encode_reverse_bit_tree(
                     self.get_dist_special(dist_slot as usize - DIST_MODEL_START),
                     dist_reduced,
                 )?;
             } else {
-                rc.encode_direct_bits(dist_reduced >> ALIGN_BITS, footer_bits - ALIGN_BITS as u32)?;
-                rc.encode_reverse_bit_tree(&mut self.dist_align, dist_reduced & ALIGN_MASK as u32)?;
+                rc.encode_direct_bits(dist_reduced >> ALIGN_BITS, footer_bits - ALIGN_BITS as u64)?;
+                rc.encode_reverse_bit_tree(&mut self.dist_align, dist_reduced & ALIGN_MASK as u64)?;
                 self.data.align_price_count -= 1;
             }
         }
@@ -380,7 +380,7 @@ impl LZMAEncoder {
         self.reps[3] = self.reps[2];
         self.reps[2] = self.reps[1];
         self.reps[1] = self.reps[0];
-        self.reps[0] = dist as i32;
+        self.reps[0] = dist as i64;
 
         self.data.dist_price_count -= 1;
         Ok(())
@@ -388,9 +388,9 @@ impl LZMAEncoder {
 
     fn encode_rep_match<W: Write>(
         &mut self,
-        rep: u32,
-        len: u32,
-        pos_state: u32,
+        rep: u64,
+        len: u64,
+        pos_state: u64,
         rc: &mut RangeEncoder<W>,
     ) -> crate::io::write_result!(W, ()) {
         if rep == 0 {
@@ -451,31 +451,31 @@ impl LZMAEncoder {
     }
 
     pub(super) fn skip(&mut self, len: usize) {
-        self.data.read_ahead += len as i32;
+        self.data.read_ahead += len as i64;
         self.lz.skip(len)
     }
 
-    pub(super) fn get_any_match_price(&self, state: &State, pos_state: u32) -> u32 {
+    pub(super) fn get_any_match_price(&self, state: &State, pos_state: u64) -> u64 {
         RangeEncoder::get_bit_price(
             self.is_match[state.get() as usize][pos_state as usize] as _,
             1,
         )
     }
 
-    pub(super) fn get_normal_match_price(&self, any_match_price: u32, state: &State) -> u32 {
+    pub(super) fn get_normal_match_price(&self, any_match_price: u64, state: &State) -> u64 {
         any_match_price + RangeEncoder::get_bit_price(self.is_rep[state.get() as usize] as _, 0)
     }
 
-    pub(super) fn get_any_rep_price(&self, any_match_price: u32, state: &State) -> u32 {
+    pub(super) fn get_any_rep_price(&self, any_match_price: u64, state: &State) -> u64 {
         any_match_price + RangeEncoder::get_bit_price(self.is_rep[state.get() as usize] as _, 1)
     }
 
     pub(super) fn get_short_rep_price(
         &self,
-        any_rep_price: u32,
+        any_rep_price: u64,
         state: &State,
-        pos_state: u32,
-    ) -> u32 {
+        pos_state: u64,
+    ) -> u64 {
         any_rep_price
             + RangeEncoder::get_bit_price(self.is_rep0[state.get() as usize] as _, 0)
             + RangeEncoder::get_bit_price(
@@ -486,11 +486,11 @@ impl LZMAEncoder {
 
     pub(super) fn get_long_rep_price(
         &self,
-        any_rep_price: u32,
-        rep: u32,
+        any_rep_price: u64,
+        rep: u64,
         state: &State,
-        pos_state: u32,
-    ) -> u32 {
+        pos_state: u64,
+    ) -> u64 {
         let mut price = any_rep_price;
 
         if rep == 0 {
@@ -508,7 +508,7 @@ impl LZMAEncoder {
                 price += RangeEncoder::get_bit_price(self.is_rep1[state.get() as usize] as _, 1)
                     + RangeEncoder::get_bit_price(
                         self.is_rep2[state.get() as usize] as _,
-                        rep as i32 - 2,
+                        rep as i64 - 2,
                     );
             }
         }
@@ -518,11 +518,11 @@ impl LZMAEncoder {
 
     pub(super) fn get_long_rep_and_len_price(
         &self,
-        rep: u32,
-        len: u32,
+        rep: u64,
+        len: u64,
         state: &State,
-        pos_state: u32,
-    ) -> u32 {
+        pos_state: u64,
+    ) -> u64 {
         let any_match_price = self.get_any_match_price(state, pos_state);
         let any_rep_price = self.get_any_rep_price(any_match_price, state);
         let long_rep_price = self.get_long_rep_price(any_rep_price, rep, state, pos_state);
@@ -531,23 +531,23 @@ impl LZMAEncoder {
 
     pub(super) fn get_match_and_len_price(
         &self,
-        normal_match_price: u32,
-        dist: u32,
-        len: u32,
-        pos_state: u32,
-    ) -> u32 {
+        normal_match_price: u64,
+        dist: u64,
+        len: u64,
+        pos_state: u64,
+    ) -> u64 {
         let mut price =
             normal_match_price + self.match_len_encoder.get_price(len as _, pos_state as _);
         let dist_state = get_dist_state(len);
 
-        if dist < FULL_DISTANCES as u32 {
+        if dist < FULL_DISTANCES as u64 {
             price += self.data.full_dist_prices[dist_state as usize][dist as usize];
         } else {
             // Note that distSlotPrices includes also
             // the price of direct bits.
             let dist_slot = LZMAEncoder::get_dist_slot(dist);
             price += self.data.dist_slot_prices[dist_state as usize][dist_slot as usize]
-                + self.data.align_prices[(dist & ALIGN_MASK as u32) as usize];
+                + self.data.align_prices[(dist & ALIGN_MASK as u64) as usize];
         }
 
         price
@@ -561,12 +561,12 @@ impl LZMAEncoder {
                 self.data.dist_slot_prices[dist_state][dist_slot] =
                     RangeEncoder::get_bit_tree_price(
                         &mut self.dist_slots[dist_state],
-                        dist_slot as u32,
+                        dist_slot as u64,
                     );
             }
 
-            for dist_slot in DIST_MODEL_END as u32..self.data.dist_slot_prices_size {
-                let count = (dist_slot >> 1) - 1 - ALIGN_BITS as u32;
+            for dist_slot in DIST_MODEL_END as u64..self.data.dist_slot_prices_size {
+                let count = (dist_slot >> 1) - 1 - ALIGN_BITS as u64;
                 self.data.dist_slot_prices[dist_state][dist_slot as usize] +=
                     RangeEncoder::get_direct_bits_price(count);
             }
@@ -587,7 +587,7 @@ impl LZMAEncoder {
                 let dist_reduced = dist - base;
                 let price = RangeEncoder::get_reverse_bit_tree_price(
                     self.get_dist_special(dist_slot - DIST_MODEL_START),
-                    dist_reduced as u32,
+                    dist_reduced as u64,
                 );
 
                 for dist_state in 0..DIST_STATES {
@@ -601,11 +601,11 @@ impl LZMAEncoder {
         assert!(dist == FULL_DISTANCES);
     }
     fn update_align_prices(&mut self) {
-        self.data.align_price_count = ALIGN_PRICE_UPDATE_INTERVAL as i32;
+        self.data.align_price_count = ALIGN_PRICE_UPDATE_INTERVAL as i64;
 
         for i in 0..ALIGN_SIZE {
             self.data.align_prices[i] =
-                RangeEncoder::get_reverse_bit_tree_price(&mut self.dist_align, i as u32);
+                RangeEncoder::get_reverse_bit_tree_price(&mut self.dist_align, i as u64);
         }
     }
 
@@ -666,7 +666,7 @@ struct LiteralSubencoder {
 }
 
 impl LiteralEncoder {
-    pub fn new(lc: u32, lp: u32) -> Self {
+    pub fn new(lc: u64, lp: u64) -> Self {
         Self {
             coder: LiteralCoder::new(lc, lp),
             subencoders: vec![LiteralSubencoder::new(); 1 << (lc + lp)],
@@ -700,7 +700,7 @@ impl LiteralEncoder {
         assert!(data.read_ahead >= 0);
         let i = self.coder.get_sub_coder_index(
             lz.get_byte_backward(1 + data.read_ahead) as _,
-            (lz.get_pos() - data.read_ahead) as u32,
+            (lz.get_pos() - data.read_ahead) as u64,
         );
         match self.subencoders[i as usize].encode(lz, data, coder, rc) {
             Ok(_) => Ok(()),
@@ -714,12 +714,12 @@ impl LiteralEncoder {
     pub(super) fn get_price(
         &self,
         encoder: &LZMAEncoder,
-        cur_byte: u32,
-        match_byte: u32,
-        prev_byte: u32,
-        pos: u32,
+        cur_byte: u64,
+        match_byte: u64,
+        prev_byte: u64,
+        pos: u64,
         state: &State,
-    ) -> u32 {
+    ) -> u64 {
         let mut price = RangeEncoder::get_bit_price(
             encoder.is_match[state.get() as usize][(pos & encoder.pos_mask) as usize] as _,
             0,
@@ -752,7 +752,7 @@ impl LiteralSubencoder {
         coder: &mut LZMACoder,
         rc: &mut RangeEncoder<W>,
     ) -> crate::io::write_result!(W, ()) {
-        let mut symbol = lz.get_byte_backward(data.read_ahead) as u32 | 0x100;
+        let mut symbol = lz.get_byte_backward(data.read_ahead) as u64 | 0x100;
 
         if coder.state.is_literal() {
             let mut subencoder_index;
@@ -768,7 +768,7 @@ impl LiteralSubencoder {
                 }
             }
         } else {
-            let mut match_byte = lz.get_byte_backward(coder.reps[0] + 1 + data.read_ahead) as u32;
+            let mut match_byte = lz.get_byte_backward(coder.reps[0] + 1 + data.read_ahead) as u64;
             let mut offset = 0x100;
             let mut subencoder_index;
             let mut match_bit;
@@ -792,8 +792,8 @@ impl LiteralSubencoder {
         Ok(())
     }
 
-    fn get_normal_price(&self, symbol: u32) -> u32 {
-        let mut price: u32 = 0;
+    fn get_normal_price(&self, symbol: u64) -> u64 {
+        let mut price: u64 = 0;
         let mut subencoder_index;
         let mut bit;
         let mut symbol = symbol | 0x100;
@@ -812,7 +812,7 @@ impl LiteralSubencoder {
         price
     }
 
-    fn get_matched_price(&self, symbol: u32, mut match_byte: u32) -> u32 {
+    fn get_matched_price(&self, symbol: u64, mut match_byte: u64) -> u64 {
         let mut price = 0;
         let mut offset = 0x100;
         let mut subencoder_index;
@@ -840,12 +840,12 @@ impl LiteralSubencoder {
 
 pub(super) struct LengthEncoder {
     coder: LengthCoder,
-    counters: crate::Vec<i32>,
-    prices: crate::Vec<crate::Vec<u32>>,
+    counters: crate::Vec<i64>,
+    prices: crate::Vec<crate::Vec<u64>>,
 }
 
 impl LengthEncoder {
-    pub fn new(pb: u32, nice_len: usize) -> Self {
+    pub fn new(pb: u64, nice_len: usize) -> Self {
         let pos_states = 1usize << pb;
         let counters = vec![0; pos_states];
         let len_symbols = (nice_len - MATCH_LEN_MIN + 1).max(LOW_SYMBOLS + MID_SYMBOLS);
@@ -864,8 +864,8 @@ impl LengthEncoder {
 
     fn encode<W: Write>(
         &mut self,
-        len: u32,
-        pos_state: u32,
+        len: u64,
+        pos_state: u64,
         rc: &mut RangeEncoder<W>,
     ) -> crate::io::write_result!(W, ()) {
         let mut len = len as usize - MATCH_LEN_MIN;
@@ -887,7 +887,7 @@ impl LengthEncoder {
         Ok(())
     }
 
-    pub fn get_price(&self, len: usize, pos_state: usize) -> u32 {
+    pub fn get_price(&self, len: usize, pos_state: usize) -> u64 {
         self.prices[pos_state][len - MATCH_LEN_MIN]
     }
 
@@ -915,7 +915,7 @@ impl LengthEncoder {
                 + choice1_price
                 + RangeEncoder::get_bit_tree_price(
                     &mut self.coder.mid[pos_state],
-                    (i - start) as u32,
+                    (i - start) as u64,
                 );
         }
         start = LOW_SYMBOLS + MID_SYMBOLS;
@@ -923,7 +923,7 @@ impl LengthEncoder {
         for i in start..self.prices[pos_state].len() {
             self.prices[pos_state][i] = choice0_price
                 + choice1_price
-                + RangeEncoder::get_bit_tree_price(&mut self.coder.high, (i - start) as u32)
+                + RangeEncoder::get_bit_tree_price(&mut self.coder.high, (i - start) as u64)
         }
     }
     //
